% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addconstraint.R
\name{addconstraint}
\alias{addconstraint}
\title{Adds constraints to the model.}
\usage{
addconstraint(x, ct)
}
\arguments{
\item{x}{A HMM object.}

\item{ct}{The additional constraints, which can be either a
matrix or a vector (see details).}
}
\value{
A HMM object similar to the input but with the additional
constraints.
}
\description{
The specified constraints are added to the model.
If parameter \code{ct} is a vector, it is expected to be a set of transition
probabilities indexed as in field transitions of the model. In this
case the constraint added is the equality between the referred probabilities of
transition.
If parameter \code{ct} is a matrix, it is expected to be a system of additional
linear equalities that the model must fulfill. Thus, the new equations
are added to the field constraints of the model.
While it is possible to use a matrix to add equality constraints, it is not
recommended because of performance.
Previous constraints of the model are preserved.
}
\examples{
model <- HMM(3)
model <- addtransition(model, c(1,2))
model <- addtransition(model, c(2,3))
model <- addtransition(model, c(3,1))
transitions(model)
constraints(model)
model <- addconstraint(model,c(2,4,5))
constraints(model)

}
\seealso{
\link{HMM}, \link{addtransition}
}
