% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtransition.R
\name{addtransition}
\alias{addtransition}
\title{Adds a transition to the model.}
\usage{
addtransition(x, t)
}
\arguments{
\item{x}{A HMM object.}

\item{t}{The transition, as a two dimensional integer vector.
The first element is the number of the initial state and the
second one the number of the final state.}
}
\value{
A HMM object similar to the input but with the additional
transition.
}
\description{
The specified transition is added to the model as a
transition with non zero probability.
Since the transition probabilities from the initial state of the
newly specified transition still have to sum up to one, the
correspondent constraint is modified accordingly.
It is not recommended to use this function to define a big model, as
it is much slower than specifying all transitions in advance.
}
\examples{
model <- HMM(3)
model <- addtransition(model, c(1,2))
model <- addtransition(model, c(2,3))
model <- addtransition(model, c(3,1))
transitions(model)
constraints(model)

}
\seealso{
\link{HMM}, \link{addconstraint}
}
