% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{constraints}
\alias{constraints}
\alias{constraints.HMM}
\title{Matrix of constraints.}
\usage{
constraints(x)

\method{constraints}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A row major sparse matrix as in \code{\link{HMM}}.
}
\description{
Returns the matrix of constraints from a HMM object.
}
\examples{
model <- HMMrectangle(3,3)
constraints(model)
nconstraints(model)
nrow(constraints(model)) # should agree

}
\seealso{
\link{HMM}, \link{nconstraints}, \link{transitions}
}
