% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEM.R
\name{createEM}
\alias{createEM}
\title{Creates the events matrix.}
\usage{
createEM(size, towers, S)
}
\arguments{
\item{size}{The number of rows and columns in the grid.}

\item{towers}{The towers(antenna) positions. This parameter is a matrix with
two rows and the number of columns equals to the number of antennas. On the
first row we have the X coordinate of the towers and on the secnd row the
Y coordinate.}

\item{S}{A function to compute the signal strength.}
}
\value{
The events matrix.
}
\description{
Creates the events matrix for a rectangular grid
according to the location of the towers and the S function.
}
