% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.R
\name{dissim}
\alias{dissim}
\alias{dissim.HMM}
\title{Computes the dissimilarity.}
\usage{
dissim(...)

\method{dissim}{HMM}(x, y1, y2)
}
\description{
This is an approximate measure of the dissimilarity
}
\keyword{internal}
