% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duptest.R
\name{duptest}
\alias{duptest}
\alias{duptest.HMM}
\title{Likelihood ratio test.}
\usage{
duptest(x, y1, y2)

\method{duptest}{HMM}(x, y1, y2)
}
\description{
The likelihood ratio test is used to check whether two sets of
observed events come from two devices carried by the same person.
}
\keyword{internal}
