% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{emissions}
\alias{emissions}
\alias{emissions.HMM}
\title{Emissions matrix}
\usage{
emissions(x)

\method{emissions}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A column major sparse matrix as in \code{\link{HMM}}.
}
\description{
Returns the matrix of emissions from a HMM object.
}
\examples{
model <- HMM(2)
emissions(model)<-diag(2)
emissions(model)

}
\seealso{
\link{HMM}, \link{emissions<-}, \link{getTM}
}
