% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{events}
\alias{events}
\title{A set of network events}
\format{
A matrix with 198 rows and 1 column
}
\usage{
data(events)
}
\description{
A dataset with network events generated by a mobile device. It is a matrix
with 1 column and 198 rows, each row corresponding to one time instant.
An element in this matrix represents the antenna ID where the mobile device
is connected at each time instant
}
\details{
A matrix with 1 column and 198 rows.
}
\keyword{datasets}
