% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fits a HMM model}
\usage{
fit(x, e, init = FALSE, method = "solnp", retrain = 1, ...)
}
\arguments{
\item{x}{A HMM model.}

\item{e}{A vector with the observed events. It admits missing values.}

\item{init}{Logical specifying whether the initial state found in x is going
to be used. Defaults to FALSE, which means that steady state initialization
will be used instead.}

\item{method}{The optimization algorithm to be used.
Defaults to \code{solnp} from package \pkg{Rsolnp}. The other possible choice
is  \code{constrOptim} from package \pkg{stats}.}

\item{retrain}{The times the optimizer will be launched with different
initial parameters. The model with higher likelihood will be returned.}

\item{...}{Arguments to be passed to the optimizer.}
}
\value{
The fitted model.
}
\description{
Fits the transition probabilities of the model by maximum likelihood.
The transition probabilities are fitted by ML, subject to the linear constraints
specified in the model. The argument retrain can be used to avoid local minima.
It is possible to specify additional non linear constraints,
passing the suitable arguments to the optimizer.
}
\examples{
model <- HMMrectangle(20,20)
S <- function(x) if (x > 5) return(0) else return(20*log(5/x))
emissions(model) <- createEM(c(20,20), towers, S)
model <- initparams(model)
model <- minparams(model)
logLik(model,events)
model <- fit(model,events)
logLik(model,events)

}
\seealso{
\link{logLik}, \link{initparams}, \link{minparams}
}
