% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstates.R
\name{fstates}
\alias{fstates}
\title{This function returns the filtered states whenever an
observation is present and predicted states otherwise.}
\usage{
fstates(P, E, e)
}
\arguments{
\item{P}{Transition matrix P(will go to i | is in j)}

\item{E}{Event observation matrix P(detection event j | is in i)}

\item{e}{Sequence of observation events. Since we are taking
time increase small, it is expected to have mostly missing
values. The first value is expected to have an observation.}
}
\value{
The filtered states.
}
\description{
This function returns the filtered states whenever an
observation is present and predicted states otherwise.
}
