% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{funique}
\alias{funique}
\title{These functions are for internal use and won't be exported
in the future.}
\usage{
funique(x)
}
\description{
These functions are for internal use and won't be exported
in the future.
}
\keyword{internal}
