% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getER.R
\name{getER}
\alias{getER}
\title{Gets the the row of the events matrix.}
\usage{
getER(point, towers, S)
}
\description{
Gets the row of the events matrix corresponding to
a point from a rectangular grid according to the location
of the towers and the S function
}
\keyword{internal}
