% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTC.R
\name{getTC}
\alias{getTC}
\title{Gets the column of the transition matrix.}
\usage{
getTC(point, size, mask)
}
\arguments{
\item{point}{We are setting the transition probabilities
starting from this point. It corresponds to the column
(point[2] - 1 ) * size[1] + point[1] of the transition matrix.}

\item{size}{Size of the grid. It is expected to be a 2 dimensional
vector, containing the number of rows and columns respectively.}

\item{mask}{Transition probabilities to the contiguous tiles. It
is expected to be a 3x3 matrix where the (2,2) element
represents the probability of staying in the current tile.}
}
\description{
Gets the column of the transition matrix corresponding to
a point from a rectangular grid according to the mask
}
