% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getTM}
\alias{getTM}
\alias{getTM.HMM}
\title{Transitions matrix getter.}
\usage{
getTM(x)

\method{getTM}{HMM}(x)
}
\arguments{
\item{x}{An HMM object.}
}
\value{
The transition matrix of the HMM.
}
\description{
Returns the transitions matrix from a HMM.
}
\details{
The HMM object contains three fields. States contains information
about the states. Transitions contains a list of matrices with
two rows. Each column represents a transition with non-zero
probability. Transitions in the same matrix have same probability.
Emissions is a matrix with the emission probabilities. These are
considered fixed.
}
\examples{
HMM(1L, list(matrix(c(1L,1L), nrow = 2)), EM = matrix(1, nrow = 1))
}
