% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{gettransmatrix}
\alias{gettransmatrix}
\alias{gettransmatrix.HMM}
\title{Transformation matrix getter}
\usage{
gettransmatrix(x)

\method{gettransmatrix}{HMM}(x)
}
\description{
Transformation matrix getter
}
