% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMrectangle.R
\name{HMMrectangle}
\alias{HMMrectangle}
\title{Basic HMM grid model.}
\usage{
HMMrectangle(x, y)
}
\arguments{
\item{x}{length of the rectangle in tiles.}

\item{y}{width of the rectangle in tiles.}
}
\value{
A HMM object.
}
\description{
Creates a basic rectangular grid model as specified.
This model is a rectangular grid where the only transitions allowed are those
between contiguous tiles. Moreover, horizontal and vertical transition
probabilities are equal for all tiles. Diagonal transition probabilities are
equal between them too, but different from the former. These constraints mean
that there are only two parameters to estimate.
The emissions field is left unassigned.
}
\examples{
model <- HMMrectangle(3,3)
nstates(model)
ntransitions(model)
nconstraints(model)

}
\seealso{
\link{emissions}, \link{minparams}
}
