% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initparams.R
\name{initparams}
\alias{initparams}
\title{Initializer for HMM objects}
\usage{
initparams(x)
}
\arguments{
\item{x}{A HMM object.}
}
\value{
An initialized HMM object.
}
\description{
Sets initial parameters for a HMM object, as specified.
The field parameters of the HMM object, which includes both
initial state and transition probabilities, is initialized at
random.

The initial states probabilities are set to an uniform
(0,1) distribution and then divided by their sum.

The initial probabilities of transition are also set to an uniform
(0,1) and in this case, projected on the constrained space. After
the projection some probability might result greater than one or
less than zero. Those probabilities are then set to uniform (0,1)
again and the process is repeated until all probabilities of
transition are in (0,1) and the constraints are satisfied.
}
\examples{
model <- HMMrectangle(3,3)
model <- initparams(model)
range(constraints(model) \%*\% c(ptransition(model), -1)) # It should be close to zero

}
\seealso{
\link{HMM}, \link{minparams}, \link{initsteady}
}
