% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsteady.R
\name{initsteady}
\alias{initsteady}
\title{Sets the initial state to the steady state}
\usage{
initsteady(x)
}
\arguments{
\item{x}{A HMM object.}
}
\value{
The same HMM object of the input with its initial state set
to steady state.
}
\description{
The initial a-priori distribution is set to the steady state of
the transition matrix.
}
\details{
The steady state is computed as the first eigenvector found that
the absolute difference of its eigenvalue and one is less than
the square root of machine epsilon.

The Markov Chain is expected to be irreducible and aperiodic. The
first because otherwise the devices would not have freedom of
movement. The second because some probabilities from one state
to itself are expected to be non zero. This implies that there
exists one unique steady state.

However, while highly encouraged, it is not enforced and it is
possible to define a non irreducible or non aperiodic model. It
is not recommendable to use initsteady with such model, as
the function only returns one of the steady states.
}
\examples{
model <- HMM(2)
model <- addtransition(model, c(1,2))
model <- addtransition(model, c(2,1))
model <- initparams(model)
istates(model)
model <- initsteady(model)
istates(model)
(istates(model) \%*\% getTM(model)) /
  sum(istates(model) \%*\% getTM(model))

}
