% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{istates<-}
\alias{istates<-}
\title{Set the initial states}
\usage{
istates(x) <- value
}
\arguments{
\item{x}{A HMM model.}

\item{value}{A numeric vector with a probability for each state that represents the
initial distribution of states.}
}
\value{
Changes the initial distribution of states in the model.
}
\description{
Sets the initial distribution of states.

The length must match the number of states, and the sum of the vector must be one.
}
\examples{
model <- HMMrectangle(3,3)
model <- initparams(model)
istates(model)
istates(model) <- (1:9) / sum(1:9)
istates(model)

}
\seealso{
\link{HMM}, \link{initparams}, \link{initsteady}, \link{fit}
}
