% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{istates}
\alias{istates}
\alias{istates.HMM}
\title{Initial state probabilities.}
\usage{
istates(x)

\method{istates}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A numeric vector with the probabilities.
}
\description{
Returns the initial state probabilities from a HMM object.
The object has to be initialized with \code{\link{initparams}},
which generates a random initial state. The vector of probabilities
follows the same order as the states, so \code{ptransition(model)[i]} is the
probability of state \code{i}. Of course, the probabilities sum up to one.
}
\examples{
model <- HMM(2)
model <- addtransition(model,c(1,2))
model <- addtransition(model,c(2,1))
model <- initparams(model)
istates(model)
sum(istates(model)) # should be one

}
\seealso{
\link{initparams}, \link{fit}, \link{nstates}, \link{initsteady}
}
