% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik.HMM}
\alias{logLik.matrix}
\title{This function returns the (minus) logLikelihood}
\usage{
logLik(...)

\method{logLik}{HMM}(x, e)

\method{logLik}{matrix}(P, E, e)
}
\description{
This function returns the minus logarithm of the likelihood to
minimize.
}
