% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minparams.R
\name{minparams}
\alias{minparams}
\title{Reparametrizes a HMM model with a minimal set of parameters.}
\usage{
minparams(x)
}
\arguments{
\item{x}{A HMM object.}
}
\value{
The same HMM object of the input with some additional
fields that store the new parameters and the matrix that
transforms these parameters in the probabilities of transition.
}
\description{
Finds a minimal set of parameters that fully determine the
probabilities of transition.

This function avoids to solve a high dimensional optimization
problem with many constraints, parametrizing the probabilities
of transition with as few parameters as possible: the number of
degrees of freedom.

A pivoted QR decomposition of the constraints matrix is done,
to get both the free parameters and the matrix that transforms
them back into the probabilities of transition.

Many constraints are expected to be equalities between two
probabilities of transition, so the function is optimized for
this special kind of constraints.
}
\examples{
model <- HMMrectangle(2,2)
model <- initparams(model)
ntransitions(model)
nconstraints(model)
model <- minparams(model)
rparams(model)
range(ptransition(model) -
  gettransmatrix(model) \%*\% c(rparams(model), 1))

}
\seealso{
\link{rparams}, \link{rparams<-}
}
