% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{nconstraints}
\alias{nconstraints}
\alias{nconstraints.HMM}
\title{Number of constraints.}
\usage{
nconstraints(x)

\method{nconstraints}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
An integer with the number of constraints of the model.
}
\description{
Returns the number of constraints from a HMM object.
}
\examples{
model <- HMM(5)
nconstraints(model)
model <- HMMrectangle(3, 3)
nconstraints(model)

}
\seealso{
\link{constraints}, \link{nstates}, \link{ntransitions}
}
