% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ntransitions}
\alias{ntransitions}
\alias{ntransitions.HMM}
\title{Number of transition parameters getter}
\usage{
ntransitions(x)

\method{ntransitions}{HMM}(x)
}
\description{
Number of transition parameters getter
}
