% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostates.R
\name{ostates}
\alias{ostates}
\title{This function returns the observed states whenever an
observation is present.}
\usage{
ostates(P, E, e)
}
\arguments{
\item{P}{Transition matrix P(will go to i | is in j)}

\item{E}{Event observation matrix P(detection event j | is in i)}

\item{e}{Sequence of observation events. Since we are taking
time increase small, it is expected to have mostly missing
values. The first value is expected to have an observation.}
}
\value{
the observed states whenever an observation is present.
}
\description{
This function returns the observed states whenever an
observation is present.
}
\keyword{internal}
