% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ptransition}
\alias{ptransition}
\alias{ptransition.HMM}
\title{Probability of transitions parameters getter}
\usage{
ptransition(x)

\method{ptransition}{HMM}(x)
}
\description{
Probability of transitions parameters getter
}
