% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{rparams<-}
\alias{rparams<-}
\title{Set reduced parameters}
\usage{
rparams(x) <- value
}
\arguments{
\item{x}{A HMM model.}

\item{value}{A numeric vector with the new parameters.}
}
\value{
Changes parameters$reducedparams$params and parameters$transitions
in the object x.
}
\description{
Sets the parameters selected by \code{minparams} function.

The function \code{minparams} selects a minimal set of parameters, that fully
determine the transition probabilities. This function sets those parameters
and recalculates all transition probabilities from them.

The model is initialized with \code{initparams} and \code{minparams} when
required.
}
\examples{
model <- HMMrectangle(3,3)
rparams(model)<-c(0.3, 0.03)
ptransition(model)

}
\seealso{
\link{minparams}, \link{rparams}, \link{initparams}
}
