% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{rparams}
\alias{rparams}
\alias{rparams.HMM}
\title{Reduced parameters.}
\usage{
rparams(x)

\method{rparams}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A numeric vector with the values of the parameters.
}
\description{
Returns the values of the minimal set of parameters.
The minimal set of parameters are selected by \code{\link{minparams}}.
They are a few probabilities of transition that determine the remaining ones
because of the constraints. They are used to fit the model.
}
\examples{
model <- HMMrectangle(3,3)
model <- initparams(model)
model <- minparams(model)
rparams(model)
ntransitions(model)
length(rparams(model)) # A much smaller parameter space!

}
\seealso{
\link{minparams}, \link{ptransition}, \link{gettransmatrix}, \link{fit}
}
