% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{setsnames<-}
\alias{setsnames<-}
\title{Set the names of the states.}
\usage{
setsnames(x) <- value
}
\arguments{
\item{x}{A HMM model.}

\item{value}{A character vector with the names.}
}
\value{
Changes states names in the object x.
}
\description{
Sets the names of the states.

The length of the character vector must match the number of states of the model.
}
\examples{
model <- HMM(3)
setsnames(model) <- c("a","b","c")
model$states$names

}
\seealso{
\link{HMM}
}
