% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstates.R
\name{sstates}
\alias{sstates}
\alias{sstates.HMM}
\alias{sstates.matrix}
\title{This function returns the smooth states}
\usage{
sstates(...)

\method{sstates}{HMM}(x, e)

\method{sstates}{matrix}(P, E, e, FS)
}
\description{
This function returns the smooth states
}
