% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileEquivalence.R
\name{tileEquivalence}
\alias{tileEquivalence}
\title{Compute equivalence between tiles and raster cells.}
\usage{
tileEquivalence(nrow, ncol, order_by = c("rasterCell"))
}
\arguments{
\item{nrow}{Number of rows in the raster.}

\item{ncol}{Number of columns in the raster.}

\item{order_by}{Either \code{rasterCell} (default) or \code{tile}.}
}
\value{
A matrix with two named columns (\code{rasterCell} and
\code{tile}) with the equivalence for each raster cell/tile.
}
\description{
This function computes the equivalence in the index of each
raster cell (according to the raster package) and each tile (according to the
network event data simulator). Raster cells are counted from left to right
starting in the upper left corner by 1. Tiles are counted from left to right
starting in the bottom left corner by 0.
}
\examples{
tileEquivalence(4, 3)

tileEquivalence(4, 3, order_by = 'tile')
}
\keyword{internal}
