% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towers.R
\docType{data}
\name{towers}
\alias{towers}
\title{Coordinates of a set of antennas/towers}
\format{
A matrix with 2 rows and 40 columns
}
\usage{
data(towers)
}
\description{
A dataset with the coordinates of a set of towers (antennas). It is matrix
with 2 rows #' and 40 columns containing the (x,y) coordinates of 40 antennas.
The first row is the x coordinates of the antennas locations and the second
row the y coordinates.
#'
}
\keyword{datasets}
