% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_postLoc.R
\name{transform_postLoc}
\alias{transform_postLoc}
\title{Transform a location probability matrix into a key-value pair.}
\usage{
transform_postLoc(
  postLocP,
  postLocJointP = NULL,
  observedValues,
  times,
  t_increment,
  ntiles,
  pad_coef = 1,
  tileEquiv.dt,
  devID,
  sparse_postLocP = FALSE,
  sparse_postLocJointP = FALSE
)
}
\arguments{
\item{postLocP}{matrix with dimensions tiles $\times$ times with the
posterior location probabilities.}

\item{observedValues}{Vector with the observed values of the event variables.}

\item{times}{vector of time instants.}

\item{t_increment}{value with the increment of time between events.}

\item{ntiles}{total number of tiles.}

\item{pad_coef}{Integer coefficient denoting the number of time periods
needed for the time padding.}

\item{tileEquiv.dt}{data.table with the equivalence between tiles and raster
cells according to the network event data simulator and rasters in R.}

\item{devID}{device ID.}

\item{postLocJointProb}{postLocP matrix of dimensions tiles $\times$ tiles * (times - 1)
with the posterior joint location probabilities.}
}
\value{
A named \code{list} with components postLocProb_HMM_deviceID.dt and
postLocJointProb with the key-value pair data.tables with these probabilities.
}
\description{
This function transforms the matrix output for posterior
location probabilities for each time instant (tiles $\times$ times) into a
key-value pair data.table with key = c(device, time, tile).

The function does also this transformation upon the matrix of joint posterior
location probabilities.
}
\keyword{internal}
