% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{transitions}
\alias{transitions}
\alias{transitions.HMM}
\title{Transitions.}
\usage{
transitions(x)

\method{transitions}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
An integer matrix with two rows as in \code{\link{HMM}}.
}
\description{
Returns the list of possible transitions from a HMM object.
Each column represents a transition, the first row is the initial state and the
second row is the final state. The transitions are ordered, first on the initial
state and then on the final state. Any transition not listed in the matrix
is supposed to be not possible (zero probability).
}
\examples{
model <- HMM(2)
transitions(model)
model <- addtransition(model,c(1,2))
model <- addtransition(model,c(2,1))
transitions(model)
}
\seealso{
\link{HMM}, \link{ntransitions}, \link{constraints}, \link{ptransition}
}
