% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi}
\alias{viterbi}
\alias{viterbi.HMM}
\title{This function returns the viterbi path.}
\usage{
viterbi(...)

\method{viterbi}{HMM}(x, e)
}
\arguments{
\item{x}{The HMM model.}

\item{e}{Sequence of observation events. Since we are taking
time increase small, it is expected to have mostly missing
values. The first value is expected to have an observation.}
}
\value{
The viterbi path.
}
\description{
This function implements the viterbi algorithm to return the
most likely path.
}
\keyword{internal}
