% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\alias{backward.HMM}
\title{The backward part of the FB algorithm.}
\usage{
backward(...)

\method{backward}{HMM}(x, y, sfactors, ...)
}
\arguments{
\item{x}{A HMM model.}

\item{y}{A vector with the observed events. It admits missing values.}

\item{sfactors}{A vector with the scale factors from the forward algorithm.}
}
\value{
A matrix that contains the (rescaled) backward probabilities.
}
\description{
Calculates the backward probabilities.
The main purpose of this function is to be combined with forward function
to calculate smooth states and smooth consecutive pairwise states. This is done
by functions \code{sstates} and \code{scpstates}.
}
\examples{
library(Matrix)
model <- HMMrectangle(3,3)
emissions(model) <- Matrix(c(1, 1, 0.5, 1, 0.5, 0, 0.5, 0, 0,
                             0, 0, 0.5, 0, 0.5, 1, 0.5, 1, 1),
                             ncol = 2, sparse = TRUE)
model <- initparams(model)
fpass <- forward(model, c(1,2,1))
backward(model, c(1,2,1), fpass$scalefactors)
}
\seealso{
\link{forward}, \link{sstates}, \link{scpstates}
}
\keyword{internal}
