% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\alias{forward.HMM}
\title{The forward part of the FB algorithm}
\usage{
forward(...)

\method{forward}{HMM}(x, y)
}
\arguments{
\item{x}{A HMM model.}

\item{y}{A vector with the observed events. It admits missing values.}
}
\value{
A list with two elements:
\itemize{
\item \emph{alpha} is a matrix that contains the filtered states.
The number of columns coincides with the number of observations,
so that column k contains the filtered state at the same time as
y[k] is observed. For missing (un)observed values, the predicted
state is returned instead.
\item \emph{scalefactors} is a vector that contains the likelihood
of each observation conditioned on all the observation from its
past. It is conveniently set to one for missing (un)observed values,
so that the joint likelihood is just the cumulative product of the
scalefactors. As obvious, its length coincides with the length of
y.
}
}
\description{
Calculates the forward probabilities.
The main purpose of this function is to be combined with backward function
to calculate smooth states and smooth consecutive pairwise states. This is done
by functions \code{sstates} and \code{scpstates}.
}
\examples{
model <- HMMrectangle(3,3)
emissions(model) <- Matrix(c(1, 1, 0.5, 1, 0.5, 0, 0.5, 0, 0,
                             0, 0, 0.5, 0, 0.5, 1, 0.5, 1, 1),
                             ncol = 2, sparse = TRUE)
model <- initparams(model)
forward(model, c(1,2,1))
}
\seealso{
\link{backward}, \link{sstates}, \link{scpstates}
}
\keyword{internal}
