% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{gettransmatrix}
\alias{gettransmatrix}
\alias{gettransmatrix.HMM}
\title{Transformation matrix}
\usage{
gettransmatrix(x)

\method{gettransmatrix}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A matrix.
}
\description{
Returns the transformation matrix that transforms the minimal
parameters into the probabilities of transition.
The transformation matrix allows obtains the probabilities of transition from
the minimal set of parameters. If we append an one at the end of the vector of
parameters, the product of this matrix by such vector is the probabilities of
transition vector.
}
\examples{
model <- HMMrectangle(3,3)
model <- initparams(model)
model <- minparams(model)
# Should be close to zero
range(ptransition(model) - gettransmatrix(model) \%*\% c(rparams(model), 1))

}
\seealso{
\link{minparams}, \link{ptransition}, \link{rparams}, \link{fit}
}
