% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik.HMM}
\title{Minus logLikelihood}
\usage{
logLik(...)

\method{logLik}{HMM}(x, e, ...)
}
\arguments{
\item{x}{A HMM model.}

\item{e}{A vector with the observed events. It admits missing values.}
}
\value{
The minus logarithm of the likelihood of the events given the model.
}
\description{
Returns the minus logarithm of the likelihood given a model and a set of observations.

A slightly modified version of the forward algorithm is used to compute the likelihood,
to avoid store unneeded data. The sign is changed because it is usual to minimize
instead maximize.
}
\examples{
model <- HMMrectangle(20,20)
S <- function(x) if (x > 5) return(0) else return(20*log(5/x))
emissions(model) <- createEM(c(20,20), towers, S)
model <- initparams(model)
model <- minparams(model)
logLik(model,events)

}
\seealso{
\link{fit}, \link{HMM}, \link{initparams}
}
