% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ntransitions}
\alias{ntransitions}
\alias{ntransitions.HMM}
\title{Number of transitions.}
\usage{
ntransitions(x)

\method{ntransitions}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
An integer with the number of possible transitions of the model.
}
\description{
Returns the number of possible transitions from a HMM object.
}
\examples{
model <- HMM(5)
ntransitions(model)
model <- addtransition(model, c(1,2))
ntransitions(model)

}
\seealso{
\link{transitions}, \link{nstates}, \link{nconstraints}
}
