% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ptransition}
\alias{ptransition}
\alias{ptransition.HMM}
\title{Probabilities of transition.}
\usage{
ptransition(x)

\method{ptransition}{HMM}(x)
}
\arguments{
\item{x}{the HMM object.}
}
\value{
A numeric vector with the probabilities of transition.
}
\description{
Returns the probabilities of transition from a HMM object.
The object has to be initialized with \code{\link{initparams}},
otherwise it will return numeric(0). The order is row major.
}
\examples{
model <- HMM(2)
model <- addtransition(model,c(1,2))
model <- addtransition(model,c(2,1))
model <- initparams(model)
ptransition(model)

}
\seealso{
\link{HMM}, \link{initparams}, \link{transitions}, \link{ntransitions}
}
